restructured the search routes
