+++ Develop/controllers/api/foodTruckRoutes.js
@@ -25,37 +25,6 @@ router.get('/:id', async (req, res) => {
     }
 })
 
-// Route to search for closest food trucks
-
-router.get('/', async (req, res) => {
-    const trucks = [];
-    try {
-        const address = req.query.address;
-        const userCoordinates = await convertAddressToCoordinates(address);
-        const truckData = await FoodTruck.findAll({ include: [MenuItem] });
-
-        // find closest foodtrucks
-        for (const truck of truckData) {
-            const truckCoordinates = await convertAddressToCoordinates(truck.address);
-            if (!truckCoordinates) {
-                continue;
-            }
-            const distance = checkDistance(userCoordinates, truckCoordinates);
-
-            if (distance < 20) {
-                trucks.push(truck);
-            };
-        };
-
-        const plainTrucks = trucks.map(truck => truck.get({ plain: true }));
-        console.log(plainTrucks);
-
-        res.render('results', { trucks: plainTrucks });
-    } catch (error) {
-        res.status(400).json(error);
-    }
-})
-
 // A route for updating a food truck - likely updating the food truck's location
 router.put('/:id', withAuth, async (req, res) => {
     try {
