+++ Develop/controllers/api/index.js
@@ -1,11 +1,11 @@
 const router = require('express').Router();
 const userRoutes = require('./userRoutes');
 const menuRoutes = require('./menuItemRoutes')
-// const commentRoutes = require('./commentRoutes');
+const searchRoute = require('./searchRoute')
 const foodTruckRoutes = require('./foodTruckRoutes');
 
 router.use('/users', userRoutes);
-// router.use('/comments', commentRoutes);
+router.use('/results', searchRoute);
 router.use('/foodTruck', foodTruckRoutes);
 router.use('/menu', menuRoutes);
