+++ Develop/views/results.handlebars
@@ -1,16 +1,15 @@
-
 <body class="user-body">
     <section class="section">
         <div class="container">
             <h1 class="title has-text-centered">Food In Your Area</h1>
             <div class="tile is-ancestor is-flex-wrap-wrap">
-                {{#each truck}}
+                {{#each trucks as |truck|}}
                 {{> tiles}}
                 {{/each}}
             </div>
         </div>
     </section>
+    <script src="/js/tileflip.js"></script>
 </body>
-</html>
 
-<script src="/js/tileflip.js"></script>
+</html>
